warn about enabling login shell in the terminal emulator
RVM does not work properly if it is run from terminal emulator that does not allow a login shell. The behaviour is easy to see by just trying to run: 
    $ rvm use

The scenario is well handled by RVM and instructions are given to the user. However, the rails install script encapsulates RVM and will never face a situation like this because, in line 16, the rvm script is being sourced.

The shell used for installation will display correctly the version numbers for Ruby and Rails, but once it exits the configuration will be lost.

I saw this behaviour in gnome-terminal, xterm and uxterm.